web-ui-fasthtml/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
httpx[http2]==0.28.1
python-dotenv==1.0.1
python-multipart==0.0.20 